@@ -33,9 +33,9 @@
     "postcss-loader": "2.0.8",
     "promise": "8.0.1",
     "raf": "3.4.0",
-    "react": "^16.8.6",
+    "react": "^16.9.0",
     "react-dev-utils": "^5.0.0",
-    "react-dom": "^16.8.6",
+    "react-dom": "^16.9.0",
     "react-redux": "^5.0.6",
     "redux": "^3.7.2",
     "style-loader": "0.19.0",
@@ -97,5 +97,9 @@
   "main": "index.js",
   "repository": "git@github.com:drewdecarme/cadent-react-redux-exercise.git",
   "author": "Drew DeCarme",
-  "license": "MIT"
+  "license": "MIT",
+  "devDependencies": {
+    "@testing-library/react": "9.1.3",
+    "redux-mock-store": "^1.5.4"
+  }
 }
