@@ -1,5 +1,3 @@
-import update from 'immutability-helper';
-
 const duckRoot = 'app/groceries/';
 
 // Constants
@@ -14,34 +12,34 @@ export const initialState = {
       id: 66,
       name: 'Bananas',
       category: 'Fruit',
-      deliveryMethod: 'Air',
+      deliveryMethod: 'Air'
     },
     {
       id: 16,
       name: 'Whole Grain Bread',
       category: 'Grains',
-      deliveryMethod: 'Air',
+      deliveryMethod: 'Air'
     },
     {
       id: 100,
       name: 'Lettuce',
       category: 'Vegitable',
-      deliveryMethod: 'Ground',
+      deliveryMethod: 'Ground'
     },
     {
       id: 10,
       name: 'Roasted Turkey',
       category: 'Deli',
-      deliveryMethod: 'Ground',
-    },
+      deliveryMethod: 'Ground'
+    }
   ],
   isItemSelected: false,
   selectedItem: {
     id: 0,
     name: '',
     category: '',
-    deliveryMethod: '',
-  },
+    deliveryMethod: ''
+  }
 };
 
 // Reducers
@@ -50,29 +48,69 @@ export default function reducer(state = initialState, action) {
 
   switch (type) {
     case ADD_ITEM:
-      return update(state, {
-        list: { $push: [payload] },
-      });
+      return {
+        ...state,
+        list: [...state.list, payload]
+      };
 
     case REMOVE_ITEM:
-      // Write a custom reducer that will remove an item from the list array
-      return state; 
+      return {
+        ...state,
+        // .filter() will remove all instances of that ID
+        // list: state.list.filter((item) => item.id !== payload)
+        list: [
+          ...state.list.slice(0, payload),
+          ...state.list.slice(payload + 1)
+        ]
+      };
 
     case SELECT_ITEM:
-      // Write a custom reducer that will select an item
-      return state;
+      return {
+        ...state,
+        isItemSelected: true,
+        selectedItem: payload
+      };
 
-    case DESELECT_ITEM:
-      // Write a customer reducer that will deselect an item
-      return state;
+    case DESELECT_ITEM: {
+      // Prevent unrelated item deselect actions from removing the selected item
+      if (payload !== state.selectedItem.id) {
+        return state;
+      }
+
+      return {
+        ...state,
+        isItemSelected: false,
+        selectedItem: {
+          id: 0,
+          name: '',
+          category: '',
+          deliveryMethod: ''
+        }
+      };
+    }
 
     default:
       return state;
   }
-};
+}
 
 // Action Creators
-export const addItem = item => ({
+export const addItem = (item) => ({
   type: ADD_ITEM,
-  payload: item,
+  payload: item
+});
+
+export const removeItem = (index) => ({
+  type: REMOVE_ITEM,
+  payload: index
+});
+
+export const selectItem = (item) => ({
+  type: SELECT_ITEM,
+  payload: item
+});
+
+export const deselectItem = (id) => ({
+  type: DESELECT_ITEM,
+  payload: id
 });
